💄 v2.1.x - Comments style conversation
Added a new display-style prop that makes it possible to display the conversation in a comment style instead of the default chat style.

When in comments mode, all messages are aligned to the left, the rounding is less pronounced (2xl to default), and all messages are gray instead of the user's messages being blue.