🐛 Send audio files mime type
The previous dev (hum hum) was sending a webm file as a mp3 file somehow to the back-end, rendering it useless on iOS and Mac OS. It worked on Android and on the web, because the webm format is supported there.

It now uses ffmpeg to transform the webm file into a valid mp3 before sending it to media for storage.